Create Binary_Search.cpp
Maximize the number of days for which P chocolates can be distributed consecutively to N people
Given an integer, P denoting the number of chocolates and an array a[] where ai denotes the type of ith chocolate. There are N people who want to eat chocolate every day. Find the maximum number of consecutive days for which N people can eat chocolates considering the following conditions:

Each of the N people must eat exactly one chocolate on a particular day.
A single person can eat chocolate of the same type only for all days.